1) Import DB from temp 2) Registering new staffs and new Sponsors Modifications

+++ database/migrations/2016_04_17_104628_create_sponsor_table.php
@@ -22,8 +22,8 @@ class CreateSponsorTable extends Migration
             $table->string('phone_no', 20)->nullable();
             $table->string('phone_no2', 20)->nullable();
             $table->text('address')->nullable();
-            $table->smallInteger('lga_id')->index()->unsigned()->nullable();
-            $table->smallInteger('salutation_id')->index()->unsigned()->nullable();
+            $table->integer('lga_id')->index()->unsigned()->nullable();
+            $table->integer('salutation_id')->index()->unsigned()->nullable();
             $table->integer('created_by', 10)->index()->unsigned();
             $table->timestamps();
         });
